Rework tagbot a lot
