+++ .github/workflows/analyze.yml
@@ -1,4 +1,4 @@
-name: Tagbot
+name: analyze
 on: [pull_request]
 
 concurrency:
@@ -25,19 +25,19 @@ jobs:
 
       - name: Install dependencies
         run: |
-          pip install gitpython requests
+          pip install gitpython
 
       - name: Run analysis
         run: |
-          python .github/workflows/tagbot.py --output tagbot-result.json
+          python .github/workflows/analyze.py --output analysis.json
         env:
           PR_NUMBER: ${{ github.event.pull_request.number }}
           PR_BASE_REF: ${{ github.event.pull_request.base.ref }}
           PR_HEAD_SHA: ${{ github.event.pull_request.head.sha }}
 
       - name: Upload analysis artifact
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a
         with:
-          name: tagbot-result
-          path: tagbot-result.json
+          name: analysis
+          path: analysis.json
           retention-days: 1
